-Opdateret BMIcalculator klassen, -Lavet form med højde og vægt til måling af BMI -Viser resultat på bmiCalculator.jsp
